Moving towards integration of google authentication
Now, the process of authentication as of yet is this -

Google button etc are loaded with login.tile.html

On sign-in the function onSignIn is called. This function extracts the idtoken
from the response and redirects to auth.php with the idtoken included in the
idtoken parameter.

Auth.php calls the Interface.Openid.workflow.php with the idtoken sent inside
the memory variable.

A new run method has been added and the old one has been renamed to run_old. The
new run loads the google-api-php-client, validates the idtoken and responds with
all the details of the user including email, first and last name etc.

Using this, do whatever the old script was doing.

Unfortunately, I am still getting an AuthenticationError. Seems like the email
id is not present in the response after validation.

Signed-off-by: Venkatesh Shukla <venkatesh.shukla.eee11@iitbhu.ac.in>